backend/requirements.txt
fastapi>=0.122.0
uvicorn[standard]>=0.38.0
sqlmodel>=0.0.27
psycopg2-binary>=2.9.11
pydantic-settings>=2.11.0
python-jose[cryptography]>=3.5.0
bcrypt>=4.2.0
python-multipart>=0.0.20
httpx>=0.28.1
email-validator>=2.2.0
